♻️ DRY refactor: Extract gold standard ingestion logic to shared module
MAJOR ARCHITECTURE IMPROVEMENT: Eliminated code duplication by extracting
the gold standard ingestion logic into a shared module.

New structure:
✅ jobs/sheets_ingestion_shared.py - GOLD STANDARD logic (parsing, ingestion, eliminations)
✅ jobs/ingest_sheet.py - Thin wrapper (66 lines) for service account auth
✅ jobs/ingest_personal_sheets.py - Thin wrapper (96 lines) for OAuth auth

Before:
❌ ingest_sheet.py: 227 lines
❌ ingest_personal_sheets.py: 200 lines
❌ Duplicated logic in both files

After:
✅ sheets_ingestion_shared.py: Single source of truth (155 lines)
✅ ingest_sheet.py: Auth wrapper only (66 lines, -161 lines!)
✅ ingest_personal_sheets.py: Auth wrapper + format adapter (96 lines, -104 lines!)

Key benefits:
- Single place to update ingestion logic
- Both methods guaranteed to use same logic
- OAuth wrapper includes format adapter to convert to raw rows
- Easier to maintain and test

Testing:
- Service account: 252 players, 1,296 picks ✅
- Week 10 verification: 30 players with picks ✅
- Week 10 breakdown matches expectations (DEN:10, SEA:6, CAR:6, BUF:4, etc.) ✅
